requirements.txt
kafka-python 
Flask 
pandas 
requests
opensky-api
psycopg2

#-e .

